requirements.txt
bs4
python-dotenv
pyppeteer
requests
tweepy
